break analysis out into a buildtool
